# subscriber-service/transport/api/openapi.yaml
openapi: 3.0.3
info:
  title: Subscriber Service
  description: Subscriber service allowing subscriptions to newsletters and post sending.
  version: 0.0.0
servers:
  - url: http://localhost
    description: Localhost server running on HTTP.
paths:
...
  "/nginx/newsletters/{newsletterId}/posts/{postId}/publish":
    get:
      summary: Send published post to subscribers
      operationId: SendPublishedPost
      parameters:
        - name: newsletterId
          in: path
          required: true
          schema:
            type: string
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post delivery process to subscribers has been successfully initiated
          description: Post sent to subscribers
        "400":
          description: Invalid request
...